Relax dependency on logstash-mixin-http_client
When we have released version 4.3.0 on this plugin we have bumped the
requirements on the mixin to be 5.1.0 or higher, the problem is this
break our contract of the 5.x of the plugin. This makes our 4.3+ plugin only
installable on Logstash 5.4+.

By looking at the changelog and commits of this plugins we did 2 major
bumps.

LS 5.1 was released with the 3.x branch
LS 5.2 was released with the 4.x branch
LS 6.x is released with the 5.x branch

When I look at the changes on the mixin we have changed how we handle
SSL verification and we have added new options for basic auth.

Theses configuration are provided by the mixin and the mixin also
provide the factory for the http client.

I think we can relax the configuration and make this plugin work with
any installed mixin the draw back of this is may make the doc out of
sync with the capabilities of this plugins. This is the problem with
shared dependencies.

Related PR: #66